refactor(canvas,core): expose one gradient-bounds function
The canvas alias only existed to keep @ripl/canvas' export surface
unchanged, which is not worth two names for one function. Renamed to
getGradientBounds and dropped the wrapper; GradientBounds follows it, so
@ripl/svg and @ripl/canvas both take the type from core.

BREAKING: @ripl/canvas no longer exports getCanvasGradientBounds or
GradientBounds. Import getGradientBounds and GradientBounds from
@ripl/core instead.
