+++ packages/svg/src/context.ts
@@ -52,6 +52,7 @@ import type {
 
 import {
     createFrameBuffer,
+    getGradientBounds,
     isGradientString,
     isPatternString,
     isTransparentColor,
@@ -59,7 +60,6 @@ import {
     parseGradient,
     parsePattern,
     radiansToDegrees,
-    resolveGradientBounds,
 } from '@ripl/core';
 
 import type {
@@ -218,7 +218,7 @@ export class SVGContext extends DOMContext<SVGSVGElement> {
         this._usedDefs.add(`gradient:${cacheKey}`);
 
         // The element's own box, not the path node's, so a multi-path element ramps once across all of them.
-        const bounds = resolveGradientBounds(this.currentRenderElement?.getBoundingBox?.(true), this.width, this.height);
+        const bounds = getGradientBounds(this.currentRenderElement?.getBoundingBox?.(true), this.width, this.height);
         const cached = this._gradientCache.get(cacheKey);
 
         if (cached) {
